Finish Typing
