--- src/index.ts
@@ -1,5 +1,0 @@
-import getStats from "./getStats";
-import fetchData from "./misc/fetchData";
-import getGame, { getAllGames } from "./misc/getGame";
-
-export { getStats, fetchData, getGame, getAllGames };
